requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==3.1.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
